move function to other file
